Fix delete so now DOM reflects change. Added some error checking as well.

+++ Phoenix/Scripts/admin.js
@@ -67,22 +67,22 @@ function createNewType(buildingCode, roomType, buildingCopy) {
     })
     .fail(function () {
         console.log("fail");
-        alert("It looks like there is no " + roomType + " RCI for the building " + copyOption + ". Are you sure you picked the right options?");
+        $("#copy-building-select").parent().append('<div class="error-message"><p>It looks like there is no ' + roomType + ' RCI for the building ' + buildingCopy + '. Are you sure you picked the right options?</p></div>');
     });
 }
 
 // Delete a certain rci type
 function deleteRciType(buildingCode, roomType) {
-    console.log("About to delete " + buildingCode + roomType);
-
+   
     $.ajax(
         {
-            method: "DELETE",
+            method: "POST",
             url: "/AdminDashboard/DeleteRciType",
             data: {buildingCode: buildingCode, roomType: roomType}
         })
     .then(function () {
-        alert("Successfully delete: " + buildingCode + roomType);
+        $("#delete-" + buildingCode + "-" + roomType).prev().remove();
+        $("#delete-" + buildingCode + "-" + roomType).remove();
     })
 }
 
@@ -106,12 +106,18 @@ $("#add-type-button").click(function () {
 
 $("#cancel-add-type").click(function () {
     $(".signature-overlay").hide();
+    $(".error-message").remove();
     $("#signature-container").toggle();
 });
 
 $("#create-new-type").click(function () {
     let buildingCode = $("#new-building-code-input").val();
     console.log(buildingCode);
+    if (buildingCode === "")
+    {
+        $("#new-building-code-input").parent().prepend('<div class="error-message"><p id="no-building-error">Please enter a building code to create a new RCI type.</p></div>');
+        return;
+    }
 
     let roomType = $("#room-type-select option:selected").text().toLocaleLowerCase();
     console.log("Room type: " + roomType);
@@ -125,12 +131,15 @@ $("#create-new-type").click(function () {
 $(".rci-types-flex").on("click", ".delete", function () {
     console.log($(this));
 
-    
     // The delete button has an id of the form "delete-WIL-common", so we split it up to get building code and room type
     let splitArray = $(this).attr("id").split("-");
     
     let buildingCode = splitArray[1];
     let roomType = splitArray[2];
 
-    deleteRciType(buildingCode, roomType);
+    if (confirm("Are you sure you wish to delete the RCI type for " + buildingCode + " - " + roomType + "?"))
+    {
+        deleteRciType(buildingCode, roomType);
+    }
+
 });
